Create the SideBar component
