--- src/components/NavBar.style.css
@@ -1,10 +1,0 @@
-.nav-menu {
-  position: fixed;
-  top: 0;
-  height: 100vh;
-  flex-direction: column;
-  background-color: red;
-}
-.nav-base {
-  z-index: 1;
-}
